+++ fabric/src/main/java/com/jahirtrap/critterarmory/init/ModLayerDefinitions.java
@@ -17,7 +17,7 @@ public class ModLayerDefinitions {
         PartDefinition partDefinition2 = partDefinition.addOrReplaceChild("head", CubeListBuilder.create().texOffs(0, 0).addBox(-2, -6, -2, 4, 6, 3, cubeDeformation), PartPose.offset(0, 15, -4));
         partDefinition2.addOrReplaceChild("beak", CubeListBuilder.create().texOffs(14, 0).addBox(-2, -4, -4, 4, 2, 2, cubeDeformation), PartPose.ZERO);
         partDefinition2.addOrReplaceChild("red_thing", CubeListBuilder.create().texOffs(14, 4).addBox(-1, -2, -3, 2, 2, 2, cubeDeformation), PartPose.ZERO);
-        partDefinition.addOrReplaceChild("body", CubeListBuilder.create().texOffs(0, 9).addBox(-3, -4, -3, 6, 8, 6, cubeDeformation), PartPose.offsetAndRotation(0, 16, 0, ((float) Math.PI / 2F), 0, 0));
+        partDefinition.addOrReplaceChild("body", CubeListBuilder.create().texOffs(0, 9).addBox(-3, -4, -3, 6, 8, 6, cubeDeformation), PartPose.offsetAndRotation(0, 16, 0, ((float) Math.PI / 2), 0, 0));
         CubeListBuilder cubeListBuilder = CubeListBuilder.create().texOffs(26, 0).addBox(-1, 0, -3, 3, 5, 3, cubeDeformation);
         partDefinition.addOrReplaceChild("right_leg", cubeListBuilder, PartPose.offset(-2, 19, 1));
         partDefinition.addOrReplaceChild("left_leg", cubeListBuilder, PartPose.offset(1, 19, 1));
@@ -30,7 +30,7 @@ public class ModLayerDefinitions {
         MeshDefinition meshDefinition = new MeshDefinition();
         PartDefinition partDefinition = meshDefinition.getRoot();
         partDefinition.addOrReplaceChild("head", CubeListBuilder.create().texOffs(0, 0).addBox(-4, -4, -6, 8, 8, 6, cubeDeformation).texOffs(22, 0).addBox("right_horn", -5, -5, -4, 1, 3, 1, cubeDeformation).texOffs(22, 0).addBox("left_horn", 4, -5, -4, 1, 3, 1, cubeDeformation), PartPose.offset(0, 4, -8));
-        partDefinition.addOrReplaceChild("body", CubeListBuilder.create().texOffs(18, 4).addBox(-6, -10, -7, 12, 18, 10, cubeDeformation).texOffs(52, 0).addBox(-2, 2, -8, 4, 6, 1, cubeDeformation), PartPose.offsetAndRotation(0, 5, 2, ((float) Math.PI / 2F), 0, 0));
+        partDefinition.addOrReplaceChild("body", CubeListBuilder.create().texOffs(18, 4).addBox(-6, -10, -7, 12, 18, 10, cubeDeformation).texOffs(52, 0).addBox(-2, 2, -8, 4, 6, 1, cubeDeformation), PartPose.offsetAndRotation(0, 5, 2, ((float) Math.PI / 2), 0, 0));
         CubeListBuilder cubeListBuilder = CubeListBuilder.create().texOffs(0, 16).addBox(-2, 0, -2, 4, 12, 4, cubeDeformation);
         partDefinition.addOrReplaceChild("right_hind_leg", cubeListBuilder, PartPose.offset(-4, 12, 7));
         partDefinition.addOrReplaceChild("left_hind_leg", cubeListBuilder, PartPose.offset(4, 12, 7));
@@ -47,7 +47,7 @@ public class ModLayerDefinitions {
         MeshDefinition meshDefinition = QuadrupedModel.createBodyMesh(12, cubeDeformation);
         PartDefinition partDefinition = meshDefinition.getRoot();
         partDefinition.addOrReplaceChild("head", CubeListBuilder.create().texOffs(0, 0).addBox(-3, -4, -6, 6, 6, 8, cubeDeformation), PartPose.offset(0, 6, -8));
-        partDefinition.addOrReplaceChild("body", CubeListBuilder.create().texOffs(28, 8).addBox(-4, -10, -7, 8, 16, 6, cubeDeformation), PartPose.offsetAndRotation(0, 5, 2, ((float) Math.PI / 2F), 0, 0));
+        partDefinition.addOrReplaceChild("body", CubeListBuilder.create().texOffs(28, 8).addBox(-4, -10, -7, 8, 16, 6, cubeDeformation), PartPose.offsetAndRotation(0, 5, 2, ((float) Math.PI / 2), 0, 0));
         return LayerDefinition.create(meshDefinition, 64, 32);
     }
 }